move cache
