+++ adana-cache-command/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "adana-cache-command"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+homepage.workspace = true
+repository.workspace = true
+readme.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+adana-db.workspace = true
+strum.workspace = true
+nom.workspace = true
+regex.workspace = true
+anyhow.workspace = true
+adana-script.workspace = true
+adana-script-core.workspace = true
+bincode.workspace = true
+nu-ansi-term.workspace = true
+arboard.workspace = true
+dirs.workspace = true
+serde_json.workspace = true
+log.workspace = true
+serde.workspace = true
