+++ adana-cache-command/src/process.rs
@@ -1,14 +1,12 @@
 use std::{collections::BTreeMap, path::PathBuf};
 
-use crate::{
-    prelude::colors::*,
-    reserved_keywords::{check_reserved_keyword, CACHE_COMMAND_DOC},
-};
+use crate::reserved_keywords::{check_reserved_keyword, CACHE_COMMAND_DOC};
 use adana_db::{DbOp, SCRIPT_CACHE_KEY};
 use adana_script::print_ast;
 use adana_script_core::primitive::RefPrimitive;
 use anyhow::Context;
 use nom::error::ErrorKind;
+use nu_ansi_term::Color::*;
 use regex::Regex;
 
 use super::{
