+++ adana-script/Cargo.toml
@@ -7,14 +7,14 @@ description.workspace = true
 authors.workspace = true
 homepage.workspace = true
 repository.workspace = true
-rustc-version.workspace = true
+rust-version.workspace = true
 
 [dependencies]
 #adana-script-core = { git = "https://github.com/nbittich/adana-script-core.git", branch = "feature/bitwise-and-new-ints" }
 nom.workspace = true
 adana-script-core.workspace = true
 anyhow.workspace = true
-slab_tree.workpace = true
+slab_tree.workspace = true
 
 [dev-dependencies]
 # some test must run sequentially.
