+++ src/prelude.rs
@@ -1,29 +1,9 @@
-pub use nom::{
-    branch::alt,
-    bytes::complete::{
-        tag, tag_no_case, take_till1, take_until, take_until1, take_while,
-        take_while1,
-    },
-    character::complete::{multispace0, multispace1, space1},
-    combinator::{cut, map, opt, rest, verify},
-    multi::{many0, separated_list0},
-    sequence::{delimited, pair, preceded, separated_pair, terminated},
-    IResult, Parser,
-};
-
-pub use log::*;
-
-pub use std::process::{Command, Stdio};
-
-pub type Res<'a, T> = IResult<&'a str, T>;
-
-pub use std::{collections::BTreeMap, fs::File, io::BufReader, path::PathBuf};
+pub use std::{collections::BTreeMap, path::PathBuf};
 
 pub mod colors {
     pub use nu_ansi_term::Color::*;
     pub use nu_ansi_term::Style;
 }
-pub use serde::{Deserialize, Serialize};
 
 pub const SHARED_LIB_DIR: &str = ".libs_adana";
